chore: remove consumer pattern
