+++ src/commerce-extensibility-1/actions/stock/commerce/actions.config.yaml
@@ -1,13 +1,3 @@
-consumer:
-  function: consumer/index.js
-  web: 'no'
-  runtime: nodejs:24
-  inputs:
-    LOG_LEVEL: debug
-    EVENT_PREFIX: $EVENT_PREFIX
-  annotations:
-    require-adobe-auth: true
-    final: true
 updated:
   function: updated/index.js
   web: 'no'
