+++ src/commerce-extensibility-1/actions/stock/external/actions.config.yaml
@@ -1,12 +1,3 @@
-consumer:
-  function: consumer/index.js
-  web: 'no'
-  runtime: nodejs:24
-  inputs:
-    LOG_LEVEL: debug
-  annotations:
-    require-adobe-auth: true
-    final: true
 updated:
   function: updated/index.js
   web: 'no'
