--- test/actions/openwhisk.test.js
@@ -1,49 +1,0 @@
-/*
-Copyright 2022 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-jest.mock("openwhisk");
-const openwhisk = require("openwhisk");
-const Openwhisk = require("../../actions/openwhisk");
-
-afterEach(() => {
-  jest.clearAllMocks();
-  jest.resetModules();
-});
-
-describe("Given Openwhisk class", () => {
-  describe("When action is invoked", () => {
-    test("then returns action response", async () => {
-      const expectedResponse = {
-        response: {
-          result: {
-            statusCode: 200,
-            body: {
-              action: "test",
-              success: true,
-            },
-          },
-        },
-      };
-
-      openwhisk.mockReturnValue({
-        actions: {
-          invoke: jest.fn().mockResolvedValue(expectedResponse),
-        },
-      });
-
-      const client = new Openwhisk("API_HOST", "API_AUTH");
-      const response = await client.invokeAction("test", {});
-
-      expect(response).toEqual(expectedResponse);
-    });
-  });
-});
